test(fee-template): add unit tests for CreateAsync , DeleteAsync and UpdateAsync with valid and invalid scenarios

+++ StudentCourseManagement.Business/Services/FinancialModule/FeeTemplateService.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
+using StudentCourseManagement.Business.Interfaces.Repositories;
 using StudentCourseManagement.Business.Interfaces.Repositories.FinancialModule;
 using StudentCourseManagement.Business.Interfaces.Services.FinancialModule;
 using StudentCourseManagement.Domain.Entities.FinancialModule;
@@ -9,22 +10,36 @@ namespace StudentCourseManagement.Business.Services.FinancialModule
     {
         private readonly IFeeTemplateRepository _feeTemplateRepository;
         private readonly ILogger<FeeTemplateService> _logger;
+        private readonly ICourseRepository _courseRepository;
 
-        public FeeTemplateService(IFeeTemplateRepository feeTemplateRepository, ILogger<FeeTemplateService> logger)
+        public FeeTemplateService(IFeeTemplateRepository feeTemplateRepository, ILogger<FeeTemplateService> logger, ICourseRepository courseRepository)
         {
             this._feeTemplateRepository = feeTemplateRepository;
             this._logger = logger;
+            this._courseRepository = courseRepository;
         }
 
         #region CURD Operation
         public async Task<bool> CreateAsync(FeeTemplate feeTemplate)
         {
+            var course = await _courseRepository.GetByIdAsync(feeTemplate.CourseId);
+            if (course == null)
+            {
+                _logger.LogWarning($"Course with Id {feeTemplate.CourseId} is inactive or not found");
+                return false;
+            }
             await _feeTemplateRepository.AddAsync(feeTemplate);
             return true;
         }
 
         public async Task<bool> DeleteAsync(int feeTemplateId)
         {
+            var feeTemplate = await _feeTemplateRepository.GetByIdAsync(feeTemplateId);
+            if (feeTemplate == null)
+            {
+                _logger.LogWarning($"FeeTemplate with Id {feeTemplateId} not found");
+                return false;
+            }
             return await _feeTemplateRepository.DeleteAsync(feeTemplateId);
         }
 
@@ -45,7 +60,19 @@ namespace StudentCourseManagement.Business.Services.FinancialModule
                 _logger.LogWarning("Id MisMatched");
                 return false;
             }
+            var existingCourse = await _courseRepository.GetByIdAsync(feeTemplate.CourseId);
+            if (existingCourse == null)
+            {
+                _logger.LogWarning($"Course with Id {feeTemplate.CourseId} not found");
+                return false;
 
+            }
+            var exisitngFeetemplate = await _feeTemplateRepository.GetByIdAsync(feeTemplateId);
+            if (exisitngFeetemplate == null)
+            {
+                _logger.LogWarning($"FeeTemplate with Id {feeTemplateId} not found");
+                return false;
+            }
             return await _feeTemplateRepository.UpdateAsync(feeTemplateId, feeTemplate);
         }
